Check available parallel resources only once
This merge changes `run.serial` and `parallel` so we only poll
for slrum resources once and use that information for all test
cases and steps.

This merge also removes the error message when you run on a login
node.  Instead, a login node is allocated a small number of cores
(4 by default) and isn't allowed to run MPI jobs (it only supports
one task).

To accommodate this, the `Step.constrain_resources()` method takes
a dictionary of available resources instead of just the number of
available cores.