Change how dynamic ntasks and min_tasks are computed
It turns out to be useful to have a method, `compute_cell_count()`
to compute the number of cells (approximate at setup and exact
at runtime) in a mesh.  If the number of cells is controlled by
config options that a user may have changed between setup and
runtime, the method should be able to recompute the correct
updated number of cells.  Or the number of cells can be determined
by a heuristic at setup but read from a file at runtime (when
the file must exist).