Compute nx, ny, dx in initial_state
Instead of computing these properties in the test case and
storing them in config options, they are computed using a
framework-level function at setup in forward (to get the cell
count) and then again at runtime in `initial_state` to make the
mesh.  Initial state stores `nx` and `ny` as attributes in the
initial state output file.  The RPE analysis step then reads
the attributes from the initial state file rather than config
options.

The reason for this change is that users should not modify these
config options.  But if a user does modify the domain size (`lx`
and `ly`) either before setup or before running, this should result
in changes in `nx` and `ny`.  There is simply no good reason for
`nx` and `ny` to be config options other than the convenience of
passing them between steps, which can be accomplished more cleanly
with file attributes.

The RPE analysis step previously used hard-coded values for the
domain extent and tick marks.  These have been updated to use
`lx` and `ly` (though that is still not entirely accurate).