Clean up cosine bell test
We want to explicitly give the package path and filename for
the config file to be read on init.  We also want to make sure
to call the superclass' version of the configure() method even
when it's not strictly necessary.  Finally, we can get rid of
some code for translating config options to namelist options that
are not actually used.