Remove `manufactured_soluiton` test case from suites
I think we want to remove this test from our suites for now since
it is failing "on purpose" (i.e. we know the MPAS-Ocean code needs
to be fixed to allows the solution to converge at the desired
rate) but this is confusing for those who are trying to use these
suites to validate unrelated code.