Rename TestCase --> Task
To accommodate more generalization of the types of tasks that
polaris is planned for, we want a more generic name for the tasks
that polaris can run on their own.

This merge renames "test cases" to "tasks" in most cases and
"test suites" to just "suites".

We leave "test groups" for now because the follow-on plan is to
remove them entirely.