Remove test groups
In order to increase the flexibility in Polaris, we have removed
test groups from the framework.  Tasks can still be organized by
a common theme in the spirit of test groups but this is no longer
a requirement.

The constructors for test groups have been replaced by simple
functions for adding the tasks directly to each component.
The subdirectories for the test groups have been prepended to
the subdirectories for the tasks instead.