Update Component, Task and Step for shared config
This merge adds a list of `configs` to the component and a
method for adding new ones (or checking for consistency if the
config has already been added).

It also adds `set_shared_config()` methods to both `Task` and
`Step` that can be used to set a shared config parser, possibly
with a local symlink.