Fix `steps_to_run` for shared config files
With this merge, each task that uses a shared config file used
by a task to be set up has its configure() method called and
then `steps_to_run` is added to the shared config options.  This
prevents a situation where `steps_to_run` is missing when a
task is set up later that overwrites the shared config file.