Allow multiple tasks to be set up with `-t` flag
This merge changes the `-t` or `--tasks` flag to `polaris setup`
to allow multiple task to be specified.

Since there is not a sensible way to specify cached steps for
each of several tasks, the `--cached` flag only works when a
single task is specified to to `-t` or `-n`.