Compute transects on quads
The code added here is similar to functionality in MPAS-Tools but
it works with data on quads, rather than triangles.  Visualization
on triangles has proven to have a number of major drawbacks. It
is not very intuitive, functionality for matplotlib functions
and methods on triangles have fewer features and are less robust
than their counterparts for quads, and there does not seem to be
support for labeled contours on triangles.