Update baroclinic channel transect plots
Color the start and end axes of the transects (now off by default).

Use the median to compute the x location of the transect.

Use vertices to determine the start and end of the transect in y
(so the transects doesn't start in the middle of a cell).

Use the min and max of temperature to determine `vmin` and `vmax`,
rather than hard-coding.