Fix planar plots with log norm
We can't pass `vmin` and `vmax` along with `norm` that also
tried to define the min and max:
```
ValueError: Passing a Normalize instance simultaneously with vmin/vmax is not supported.  Please pass vmin/vmax directly to the norm when creating it.
```

This merge also drops an attempt to keep `vmin` positive for log
norm, instead requiring the calling code to set reasonable bounds
for log plots.