Make mapping filename an absolute path
In the `MappingFileStep`, it is important that we make not just
the scrip files but also the mapping file itself have an absolute
path so other steps that have the mapping step as a dependency can
find the mapping file.  Otherwise, they unnecessarily recreated
it (and do so without proper logging).