Add `define_smoothing()` method to `RemapTopoStep`
This method should be overridden by descending classes to define
a spatially variable smoothing field on the MPAS base mesh.

For a bit more clarity, the module containing `RemapTopoStep` is
renamed from `step` to `remap`.

This merge also adds the masking step as a prerequisite of this step.
It revamps the renaming of remapped variables to fit better with
anticipated downstream needs.

Finally, there is no longer any attempt to rename or add to the
remapped variables.  Instead, that will be left to steps that
follow (e.g. after culling).