By default, detect ocean model based on files in path
When calling `polaris setup` or `polaris suite` without the
`--model` flag, the default behavior will now be to look for
a list of required files in the component's path (providied
with the `-p` flag) to see which model is detecte.  If no model
can be detected in this way, an error is raised.